+++ eventful-postgresql/src/Eventful/Store/Postgresql.hs
@@ -30,7 +30,7 @@ postgresqlEventStore config =
   let
     getLatestVersion = sqlMaxEventVersion config maxPostgresVersionSql
     getEvents = sqlGetAggregateEvents config
-    storeEvents' = sqlStoreEvents config maxPostgresVersionSql
+    storeEvents' = sqlStoreEvents config (Just tableLockFunc) maxPostgresVersionSql
     storeEvents = transactionalExpectedWriteHelper getLatestVersion storeEvents'
   in EventStore{..}
 
@@ -46,3 +46,16 @@ initializePostgresqlEventStore pool = do
   _ <- liftIO $ runSqlPool (runMigrationSilent migrateSqlEvent) pool
 
   return ()
+
+-- | We need to lock the events table or else our global sequence number might
+-- not be monotonically increasing over time from the point of view of a
+-- reader.
+--
+-- For example, say transaction A begins to write an event and the
+-- auto-increment key is 1. Then, transaction B starts to insert an event and
+-- gets an id of 2. If transaction B is quick and completes, then a listener
+-- might see the event from B and thinks it has all the events up to a sequence
+-- number of 2. However, once A finishes and the event with the id of 1 is
+-- done, then the listener won't know that event exists.
+tableLockFunc :: Text -> Text
+tableLockFunc tableName = "LOCK " <> tableName <> " IN EXCLUSIVE MODE"