+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 * Refactored `ProcessManager` so it acts more like a read model that emits
   commands and events. Got rid of `ProcessManagerRouter`
+* Fix race condition in postgres event store where sequence numbers won't be
+  monotonically increasing when multiple writers are writing to the events
+  table. Fixed with an explicit lock on the events table.
 * Restructured record types for events paired with version numbers and global
   sequence numbers.